COMMERCE-2023 - refactor, move metadata keys resolvers to separate package
